+++ mrmustard/math/tensorflow_wrapper.py
@@ -19,6 +19,7 @@ from typing import Callable, List, Optional, Sequence, Tuple, Union
 import numpy as np
 import tensorflow as tf
 import tensorflow_probability as tfp
+import hashlib
 
 from mrmustard import settings
 from mrmustard.math.autocast import Autocast
@@ -69,7 +70,24 @@ class TFMath(MathInterface):
         return tf.convert_to_tensor(array, dtype=dtype)
 
     def atleast_1d(self, array: tf.Tensor, dtype=None) -> tf.Tensor:
-        return self.cast(tf.reshape(array, [-1]), dtype)
+        return self.cast(tf.experimental.numpy.atleast_1d(array), dtype)
+
+    def atleast_2d(self, array: tf.Tensor, dtype=None) -> tf.Tensor:
+        return self.cast(tf.experimental.numpy.atleast_2d(array), dtype)
+
+    def atleast_3d(self, array: tf.Tensor, dtype=None) -> tf.Tensor:
+        # NOTE in TF 2.14, this function adds an axis AT THE END (!)
+        array = self.atleast_2d(self.atleast_1d(array, dtype))
+        if len(array.shape) == 2:
+            array = self.expand_dims(array, 0)
+        return array
+
+    def block_diag(self, mat1: tf.Tensor, mat2: tf.Tensor) -> tf.Tensor:
+        Za = self.zeros((mat1.shape[-2], mat2.shape[-1]), dtype=mat1.dtype)
+        Zb = self.zeros((mat2.shape[-2], mat1.shape[-1]), dtype=mat1.dtype)
+        return self.concat(
+            [self.concat([mat1, Za], axis=-1), self.concat([Zb, mat2], axis=-1)], axis=-2
+        )
 
     def cast(self, array: tf.Tensor, dtype=None) -> tf.Tensor:
         if dtype is None:
@@ -165,11 +183,8 @@ class TFMath(MathInterface):
         return tf.gather(array, indices, axis=axis)
 
     def hash_tensor(self, tensor: tf.Tensor) -> int:
-        try:
-            REF = tensor.ref()
-        except AttributeError as e:
-            raise TypeError("Cannot hash tensor") from e
-        return hash(REF)
+        REF = self.asnumpy(tensor).tobytes()
+        return hashlib.sha256(REF).hexdigest()
 
     def imag(self, array: tf.Tensor) -> tf.Tensor:
         return tf.math.imag(array)
@@ -309,13 +324,13 @@ class TFMath(MathInterface):
 
     def unique_tensors(self, lst: List[Tensor]) -> List[Tensor]:
         hash_dict = {}
-        for tensor in lst:
+        for i, tensor in enumerate(lst):
             try:
                 if (hash := self.hash_tensor(tensor)) not in hash_dict:
-                    hash_dict[hash] = tensor
+                    hash_dict[hash] = (i, tensor)
             except TypeError:
                 continue
-        return list(hash_dict.values())
+        yield from hash_dict.values()
 
     def zeros(self, shape: Sequence[int], dtype=tf.float64) -> tf.Tensor:
         return tf.zeros(shape, dtype=dtype)